requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.94.0
mem0ai==1.0.11
supabase==2.10.0
httpx==0.27.0
tenacity==8.3.0
pydantic-settings==2.3.0
python-dotenv==1.0.1
structlog==24.4.0
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-httpx==0.30.0
pytest-cov==5.0.0
